meeting-analyzer-mcp/pyproject.toml
[project]
name = "meeting-analyzer-mcp"
version = "0.1.0"
description = "MCP server for analyzing meeting transcripts to determine fruitfulness"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_analyzer"]

[project.scripts]
meeting-analyzer = "meeting_analyzer.server:run"
meeting-analyzer-api = "meeting_analyzer.api:app"
